requirements.txt
plexapi==4.15.14
pypresence==4.3.0
pillow==10.3.0